requirements.txt
# Core dependencies
pyarrow>=14.0.0
duckdb>=0.9.0
polars>=0.19.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
certifi>=2023.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
click>=8.1.0
rich>=13.0.0  # For beautiful CLI output

# Reporting
jinja2>=3.1.0  # For HTML report templates

# Alerting (optional)
requests>=2.31.0  # For Slack webhooks

# Development/Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
